test(mitm-addon): split x connector usage tests
