+++ crates/runner/mitm-addon/tests/test_error_handler.py
@@ -242,134 +242,6 @@ class TestErrorHandler:
         assert "api_key=secret" not in entry["message"]
         assert "#frag" not in entry["message"]
 
-    def test_error_logs_connector_usage_for_x_stream(
-        self, tmp_path, real_flow, mitm_ctx, headers, fresh_usage_executor, usage_webhook_api
-    ):
-        """Mid-flight stream crash: partial counts still reported (issue #9534)."""
-        flow = real_flow(with_response=False, host="api.x.com", path="/2/tweets/search/stream")
-        log_path = str(tmp_path / "network.jsonl")
-        proxy_log = tmp_path / "proxy.jsonl"
-        flow.metadata["vm_run_id"] = "run-abc-123"
-        flow.metadata["vm_network_log_path"] = log_path
-        flow.metadata["vm_proxy_log_path"] = str(proxy_log)
-        flow.metadata["original_url"] = "https://api.x.com/2/tweets/search/stream"
-        flow.metadata["firewall_action"] = "ALLOW"
-        flow.metadata["firewall_name"] = "x"
-        flow.metadata["firewall_billable"] = True
-        flow.metadata["firewall_permission"] = "tweet.read"
-        flow.metadata["firewall_rule_match"] = "GET /2/tweets/search/stream"
-        flow.metadata["x_ndjson_state"] = {
-            "data_count": 23,
-            "includes": {"users": 5},
-            "lines_parsed": 23,
-            "lines_failed": 0,
-        }
-        flow.metadata["stream_buffer"] = bytearray()
-        flow.metadata["stream_buffer_state"] = {"truncated": False}
-        flow.response = tutils.tresp(status_code=200)
-        # X streams return application/json with chunked transfer, not x-ndjson.
-        flow.response.headers = header_map({"content-type": "application/json"})
-        flow.error = Error("connection reset by peer")
-        flow.metadata["vm_sandbox_token"] = "test-token"
-
-        with usage_webhook_api() as webhook:
-            mitm_addon.error(flow)
-            usage.flush_usage_events(trigger="test")
-            usage.webhook.usage_executor.shutdown(wait=True)
-
-        assert webhook.request_count > 0
-        payloads = webhook.usage_events()
-        by_cat = {p["category"]: p["quantity"] for p in payloads}
-        assert by_cat["posts.read"] == 23
-        assert by_cat["user.read"] == 5
-
-    def test_full_pipeline_stream_error_midflight(
-        self, tmp_path, real_flow, mitm_ctx, headers, fresh_usage_executor, usage_webhook_api
-    ):
-        """End-to-end: responseheaders → partial chunks → error() logs observed counts.
-
-        Simulates a real scenario: stream opens successfully, a few tweets
-        arrive, then the connection resets.  No pre-populated state — the
-        incremental parser must have accumulated counts from the chunks.
-        """
-        flow = real_flow(with_response=False, host="api.x.com", path="/2/tweets/search/stream")
-        log_path = str(tmp_path / "network.jsonl")
-        proxy_log = tmp_path / "proxy.jsonl"
-        flow.metadata["vm_run_id"] = "run-abc-123"
-        flow.metadata["vm_network_log_path"] = log_path
-        flow.metadata["vm_proxy_log_path"] = str(proxy_log)
-        flow.metadata["original_url"] = "https://api.x.com/2/tweets/search/stream"
-        flow.metadata["firewall_action"] = "ALLOW"
-        flow.metadata["firewall_name"] = "x"
-        flow.metadata["firewall_billable"] = True
-        flow.metadata["firewall_permission"] = "tweet.read"
-        flow.metadata["firewall_rule_match"] = "GET /2/tweets/search/stream"
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "application/json"})
-        )
-
-        # 1. Register parser
-        mitm_addon.responseheaders(flow)
-        callback = response_stream(flow)
-        assert "x_ndjson_state" in flow.metadata
-
-        # 2. Receive two complete tweets, then a partial third (cut off)
-        callback(b'{"data":{"id":"1"},"includes":{"users":[{"id":"u1"}]}}\n')
-        callback(b'{"data":{"id":"2"}}\n')
-        callback(b'{"data":{"id":"3"}')  # no trailing \n — connection dies here
-
-        # 3. Connection aborts
-        flow.error = Error("connection reset by peer")
-        flow.metadata["vm_sandbox_token"] = "test-token"
-
-        with usage_webhook_api() as webhook:
-            mitm_addon.error(flow)
-            usage.flush_usage_events(trigger="test")
-            usage.webhook.usage_executor.shutdown(wait=True)
-
-        # 4. Billing must reflect the 2 complete tweets (partial 3rd is dropped)
-        payloads = webhook.usage_events()
-        by_cat = {p["category"]: p["quantity"] for p in payloads}
-        assert by_cat["posts.read"] == 2  # not 3 — partial trailing dropped
-        assert by_cat["user.read"] == 1
-
-    def test_full_pipeline_stream_error_counts_complete_final_line_without_newline(
-        self, tmp_path, real_flow, mitm_ctx, headers, fresh_usage_executor, usage_webhook_api
-    ):
-        """Connection error finalizes a complete NDJSON row without trailing newline."""
-        flow = real_flow(with_response=False, host="api.x.com", path="/2/tweets/search/stream")
-        log_path = str(tmp_path / "network.jsonl")
-        proxy_log = tmp_path / "proxy.jsonl"
-        flow.metadata["vm_run_id"] = "run-abc-123"
-        flow.metadata["vm_network_log_path"] = log_path
-        flow.metadata["vm_proxy_log_path"] = str(proxy_log)
-        flow.metadata["original_url"] = "https://api.x.com/2/tweets/search/stream"
-        flow.metadata["firewall_action"] = "ALLOW"
-        flow.metadata["firewall_name"] = "x"
-        flow.metadata["firewall_billable"] = True
-        flow.metadata["firewall_permission"] = "tweet.read"
-        flow.metadata["firewall_rule_match"] = "GET /2/tweets/search/stream"
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "application/json"})
-        )
-
-        mitm_addon.responseheaders(flow)
-        callback = response_stream(flow)
-        callback(b'{"data":{"id":"1"}}\n')
-        callback(b'{"data":{"id":"2"},"includes":{"users":[{"id":"u2"}]}}')
-        flow.error = Error("connection reset by peer")
-        flow.metadata["vm_sandbox_token"] = "test-token"
-
-        with usage_webhook_api() as webhook:
-            mitm_addon.error(flow)
-            usage.flush_usage_events(trigger="test")
-            usage.webhook.usage_executor.shutdown(wait=True)
-
-        payloads = webhook.usage_events()
-        by_cat = {payload["category"]: payload["quantity"] for payload in payloads}
-        assert by_cat == {"posts.read": 2, "user.read": 1}
-        assert "connector_response_finish" not in flow.metadata
-
     def test_full_path_error_to_webhook(
         self, tmp_path, real_flow, mitm_ctx, fresh_usage_executor, usage_webhook_api
     ):
